Editmode for login brokers

+++ src/app/components/login/login.component.ts
@@ -1,6 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 import { Observable } from 'rxjs/Observable';
+import { ISubscription } from 'rxjs/Subscription';
 import { Store } from '@ngrx/store';
 
 import { IAppState } from 'app/app.reducers';
@@ -8,20 +9,26 @@ import { IAppState } from 'app/app.reducers';
 import { LoginBroker } from 'app/common/login-broker';
 import { LoginService } from 'app/services/login.service';
 import { BrokerService } from 'app/services/broker.service';
+import { DomUtil } from 'app/util/dom-util';
 
 @Component({
   selector: 'hc-login',
   templateUrl: './login.component.html',
   styleUrls: ['./login.component.scss']
 })
-export class LoginComponent implements OnInit {
+export class LoginComponent implements OnInit, OnDestroy {
 
   public brokers$: Observable<Array<LoginBroker>>;
-  public activeBrokerName$: Observable<string>;
+  public activeBrokerName: string;
   public addForm: FormGroup;
   public nameControl: FormControl;
   public urlControl: FormControl;
-  public editing: boolean;
+  public editorShown: boolean;
+  public editingIndex: number;
+
+  private brokerValidator: any;
+
+  private activeBrokerNameSub: ISubscription;
 
   constructor(
     private loginService: LoginService,
@@ -29,10 +36,14 @@ export class LoginComponent implements OnInit {
     private store: Store<IAppState>) { }
 
   public ngOnInit(): void {
-    this.brokers$ = this.loginService.getBrokers();
-    this.activeBrokerName$ = this.store.select(appState => appState.broker.activeBrokerName);
+    this.brokerValidator = this.createBrokerValidator(this.loginService);
 
-    this.nameControl = new FormControl(null, Validators.required, this.createBrokerValidator(this.loginService));
+    this.brokers$ = this.loginService.getBrokers();
+    this.activeBrokerNameSub = this.store.select(appState => appState.broker.activeBrokerName).subscribe(name => {
+      this.activeBrokerName = name;
+    });
+
+    this.nameControl = new FormControl(null);
     this.urlControl = new FormControl(null, Validators.required);
 
     this.addForm = new FormGroup({
@@ -41,30 +52,64 @@ export class LoginComponent implements OnInit {
     });
   }
 
-  public openEditor(): void {
-    this.editing = true;
+  public ngOnDestroy(): void {
+    if (this.activeBrokerNameSub) {
+      this.activeBrokerNameSub.unsubscribe();
+    }
+  }
+
+  public openEditorNew(): void {
+    this.nameControl.setValidators(Validators.required);
+    this.nameControl.setAsyncValidators(this.brokerValidator);
+    this.editingIndex = null;
+    this.editorShown = true;
+  }
+
+  public openEditorUpdate(index: number, broker: LoginBroker): void {
+    if (broker.name === this.activeBrokerName) {
+      return;
+    }
+
+    this.nameControl.setValidators(Validators.required);
+    this.nameControl.clearAsyncValidators();
+    this.editingIndex = index;
+    this.nameControl.setValue(broker.name);
+    this.urlControl.setValue(broker.url);
+    this.editorShown = true;
   }
 
   public exitEditor(): void {
     this.addForm.reset();
-    this.editing = false;
+    this.editorShown = false;
+    this.editingIndex = null;
   }
 
-  public addBroker(): void {
-    const newBroker: LoginBroker = new LoginBroker();
-    newBroker.name = this.nameControl.value;
-    newBroker.url = this.urlControl.value;
+  public saveBroker(): void {
+    const broker: LoginBroker = new LoginBroker();
+    broker.name = this.nameControl.value;
+    broker.url = this.urlControl.value;
 
-    this.loginService.addBroker(newBroker);
+    if (this.editingIndex != null && this.editingIndex >= 0) {
+      this.loginService.updateBroker(this.editingIndex, broker);
+    } else {
+      this.loginService.addBroker(broker);
+    }
+
     this.addForm.reset();
     this.exitEditor();
   }
 
-  public deleteBroker(index: number): void {
-    this.loginService.deleteBroker(index);
+  public deleteBroker(index: number, broker: LoginBroker): void {
+    if (broker.name !== this.activeBrokerName) {
+      this.loginService.deleteBroker(index);
+    }
   }
 
-  public loadBroker(broker: LoginBroker): void {
+  public loadBroker(event: any, broker: LoginBroker): void {
+    if (broker.name === this.activeBrokerName && event.target && DomUtil.isInClass(event.target, 'broker-actions')) {
+      return;
+    }
+
     this.brokerService.login(broker);
   }
